refactor(config): move classes from config.py
- Move authent classes to auth_config.py
- Move UserConfig to user_config.py
