+++ ggshield/core/config/user_config.py
@@ -1,0 +1,105 @@
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional, Set, Tuple
+
+import click
+
+from ggshield.core.config.utils import YAMLFileConfig, get_global_path, load_yaml
+from ggshield.core.constants import (
+    DEFAULT_LOCAL_CONFIG_PATH,
+    GLOBAL_CONFIG_FILENAMES,
+    LOCAL_CONFIG_PATHS,
+)
+from ggshield.core.types import IgnoredMatch
+from ggshield.core.utils import api_to_dashboard_url
+
+
+@dataclass
+class UserConfig(YAMLFileConfig):
+    """
+    Holds all ggshield settings defined by the user in the .gitguardian.yaml files
+    (local and global).
+    """
+
+    instance: Optional[str] = None
+    all_policies: bool = False
+    exit_zero: bool = False
+    matches_ignore: List[IgnoredMatch] = field(default_factory=list)
+    paths_ignore: Set[str] = field(default_factory=set)
+    show_secrets: bool = False
+    verbose: bool = False
+    allow_self_signed: bool = False
+    max_commits_for_hook: int = 50
+    banlisted_detectors: Set[str] = field(default_factory=set)
+    ignore_default_excludes: bool = False
+
+    def update_config(self, data: Dict[str, Any]) -> bool:
+        # If data contains the old "api-url" key, turn it into an "instance" key,
+        # but only if there is no "instance" key
+        try:
+            api_url = data.pop("api_url")
+        except KeyError:
+            pass
+        else:
+            if "instance" not in data:
+                data["instance"] = api_to_dashboard_url(api_url, warn=True)
+        return super(UserConfig, self).update_config(data)
+
+    def save(self, config_path: str) -> None:
+        """
+        Save config to config_path
+        """
+        self.save_yaml(config_path)
+
+    @classmethod
+    def load(cls, config_path: Optional[str] = None) -> Tuple["UserConfig", str]:
+        """
+        Load the various user configs files to create a UserConfig object:
+        - global user configuration file (in the home)
+        - local user configuration file (in the repository)
+
+        Returns a UserConfig instance, and the path where updates should be saved
+        """
+
+        user_config = UserConfig()
+        if config_path:
+            data = load_yaml(config_path) or {}
+            user_config.update_config(data)
+            return user_config, config_path
+
+        for global_config_filename in GLOBAL_CONFIG_FILENAMES:
+            global_config_path = get_global_path(global_config_filename)
+            global_data = load_yaml(global_config_path)
+            if global_data and user_config.update_config(global_data):
+                break
+
+        for local_config_path in LOCAL_CONFIG_PATHS:
+            local_data = load_yaml(local_config_path)
+            if local_data and user_config.update_config(local_data):
+                config_path = local_config_path
+                break
+
+        if config_path is None:
+            config_path = DEFAULT_LOCAL_CONFIG_PATH
+        return user_config, config_path
+
+    def add_ignored_match(self, secret: Dict) -> None:
+        """
+        Add secret to matches_ignore.
+        if it matches an ignore not in dict form, it converts it.
+        """
+        found = False
+        for i, match in enumerate(self.matches_ignore):
+            if isinstance(match, dict):
+                if match["match"] == secret["match"]:
+                    found = True
+                    # take the opportunity to name the ignored match
+                    if not match["name"]:
+                        match["name"] = secret["name"]
+            elif isinstance(match, str):
+                if match == secret["match"]:
+                    found = True
+                    self.matches_ignore[i] = secret
+            else:
+                raise click.ClickException("Wrong format found in ignored matches")
+        if not found:
+            self.matches_ignore.append(secret)
